feat(gepa): extract GEPA and delete old optimizers
GEPA and the Pareto frontier now live in dsrs-gepa. COPRO and MIPROv2 source files are deleted instead of kept behind compatibility exports, matching the hard-cutover design.

The non-obvious break was predictor discovery: the Facet walker still recognized only the old dspy_rs::predictors::predict::Predict shape. Updating that identity to dsrs_predict::predict makes GEPA usable after the split, and the extracted GEPA test now exercises state restoration through the new crate boundary.

Deleted optimizer-only coverage and examples:
- test_miprov2.rs
- test_optimizer_named_parameters_integration.rs
- test_optimizer_typed_metric.rs
- examples/02-module-iteration-and-updation.rs
- examples/04-optimize-hotpotqa.rs
- examples/08-optimize-mipro.rs
- examples/94-smoke-slice5-optimizer-interface.rs

Verification:
- cargo test -p dsrs-gepa
- cargo check --workspace
- cargo test --workspace --no-run

Scaffolding: dspy-rs still has a temporary optimizer.rs re-export for GEPA while the final tests/examples relocation and crate deletion are pending.
