+++ crates/dspy-rs/src/optimizer.rs
@@ -1,0 +1,1 @@
+pub use dsrs_gepa::*;
