+++ crates/dsrs-gepa/Cargo.toml
@@ -8,3 +8,17 @@ repository = "https://github.com/krypticmouse/DSRs"
 description = "DSRs GEPA optimizer support."
 
 [dependencies]
+anyhow = "1.0.99"
+bon = "3.7.0"
+dsrs-core = { path = "../dsrs-core" }
+dsrs-evaluate = { path = "../dsrs-evaluate" }
+dsrs-lm = { path = "../dsrs-lm" }
+dsrs-predict = { path = "../dsrs-predict" }
+facet = { git = "https://github.com/darinkishore/facet", rev = "cc8613c97cd1ec03e63659db34a947989b45c8a5", default-features = false, features = ["std"] }
+rand = "0.8.5"
+serde = { version = "1.0.219", features = ["derive"] }
+tracing = "0.1.44"
+
+[dev-dependencies]
+dsrs_macros = { version = "0.7.2", path = "../dsrs-macros" }
+tokio = { version = "1.46.1", features = ["macros", "rt"] }
